Feat/trade 거래완료버튼, 관리자 폴더, html수정

+++ src/main/java/com/o2/site/trade/dto/ApplicationDto.java
@@ -9,7 +9,7 @@ import lombok.ToString;
 @ToString
 public class ApplicationDto {
     private int tradeNo;
-    private int memberNo;
+    private Long memberNo;
     private String title;
     private String category;
     private String content;
